+++ src/pages/MatchList/components/ReservationList.tsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
 import { Link } from "react-router-dom";
-import { Content } from "../../common/types/type";
-import { TimeButton } from "../../common/components/atoms/TimeButton";
+import { Content } from "@/common/types/type";
+import { TimeButton } from "@/common/components/atoms/TimeButton";
 interface ReservationListProps {
   filterData: Content[];
 }
