+++ src/pages/MatchList/index.tsx
@@ -1,16 +1,14 @@
 import { useEffect, useState } from "react";
-import MatchListMain from "../components/matchList/MatchListMain";
-import MatchListTab from "../components/matchList/Tab";
+import MatchListMain from "@/pages/MatchList/components/MatchListMain";
+import MatchListTab from "@/pages/MatchList/components/Tab";
 import axios from "axios";
 
+import Error from "@/pages/ErrorPage";
 import { localUrl } from "@/common/constants";
 import Loading from "@/common/components/atoms/Loading";
-import MatchTabList from "@/components/matchList/MatchTabList";
-
-import { useQuery } from "@tanstack/react-query";
+import MatchDetailMenu from "@/pages/MatchList/components/MatchDetailMenu";
 import { Content } from "@/common/types/type";
-import ErrorPage from "@/pages/ErrorPage";
-
+import { useQuery } from "@tanstack/react-query";
 interface MatchListProps {
   sport: string;
 }
@@ -21,11 +19,7 @@ export default function MatchList({ sport }: MatchListProps) {
 
   //예매일정
   const [sceduleLen, setScheduleLen] = useState(0);
-  // const [matchData, setMatchData] = useState<Content[]>([]);
   const [filterData, setFilterData] = useState<Content[]>([]);
-
-  //홈구장
-  // const [stadium, setStadium] = useState<string>("");
 
   const {
     data: matchData = [],
@@ -34,7 +28,7 @@ export default function MatchList({ sport }: MatchListProps) {
   } = useQuery({
     queryKey: ["matches", sport],
     queryFn: async () => {
-      const res = await axios.get(`${localUrl}/content?teamId=${sport}`);
+      const res = await axios.get(`http://localhost:3000/content?teamId=${sport}`);
       const nowTime = new Date();
 
       return res.data.filter((game: { sportName: string; timeOffSale: string }) => {
@@ -55,18 +49,16 @@ export default function MatchList({ sport }: MatchListProps) {
     }
   }, [selectedTeam, matchData]);
 
-  // console.log('filterData :',filterData);
-
   if (isLoading) return <Loading />;
-  if (error) return <ErrorPage />;
+  if (error) return <Error />;
   return (
     <div className="flex w-[1280px] flex-row pt-10">
       <MatchListTab sport={sport} setSelectedTeam={setSelectedTeam} />
       <div className="flex w-full pl-[30px]">
         {selectedTeam === "전체 일정" ? (
-          <MatchTabList selectedTeam={selectedTeam} filterData={matchData} />
+          <MatchDetailMenu selectedTeam={selectedTeam} filterData={matchData} />
         ) : selectedTeam ? (
-          <MatchTabList selectedTeam={selectedTeam} filterData={filterData} />
+          <MatchDetailMenu selectedTeam={selectedTeam} filterData={filterData} />
         ) : (
           <MatchListMain sceduleLen={sceduleLen} sport={sport} />
         )}
